+++ CHANGELOG.md
@@ -25,8 +25,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   name`.
 - Integration of [`step-kms-plugin`](https://github.com/smallstep/step-kms-plugin)
   on `step certificate create` and `step certificate sign`.
+
+## [0.21.0] - 2022-07-06 
+### Added 
+- Device Authorization Grant flow for input constrained devices needing OAuth
+credentials. `--console-flow` flag in `step oauth` for selecting which 
+alternative OAuth flow to use.
 ### Fixed
-- Added back --domain and --remove-domain flags to provisioner CRUD
+- Added back --domain and --remove-domain flags to provisioner CRUD.
+### Removed
+- The `beta` prefix for remote provisioner and admin management.
 
 ## [0.20.0] - 2022-05-26
 ### Added